(libfreefare-desfire) Remove a brunch of -Wall -Wextra -pedantic warnings.

+++ branches/libfreefare-desfire/libfreefare/freefare.c
@@ -78,7 +78,7 @@ freefare_get_tags (nfc_device_t *device)
 	bool found = false;
 	struct supported_tag *tag_info;
 
-	for (int i = 0; i < sizeof (supported_tags) / sizeof (struct supported_tag); i++) {
+	for (size_t i = 0; i < sizeof (supported_tags) / sizeof (struct supported_tag); i++) {
 	    if ((target_info.nai.abtAtqa[0] == supported_tags[i].ATQA[0]) &&
 		(target_info.nai.abtAtqa[1] == supported_tags[i].ATQA[1]) &&
 		(target_info.nai.btSak == supported_tags[i].SAK)) {
